presenter: Ensure that height doesn't exceed screen height
Otherwise, the fullscreening may be broken in odd ways, at least in
Gnome 3. We give the current slide view up to 0.9*0.8 of the height and
the strict views up to 0.9*0.2 (minus padding) of the height, with the
status bar at the bottom getting the last 0.1.  We look at the size
request for the current view and set the width of the next view to fill
the remaining space (and hope that the allocation will be as requested).