MovieManager is aware of changing pages
This allows it to stop (and maybe delete) the currently running videos.
This required hooking up the MovieManager to a View.Pdf signal, which
wasn't pleasant given the many layers of indirection and protected
attributes.  To help, the views now have a get_main_view method that should
return the current slide view.  PresentationController gets this view for
the first controllable added and stores it in the attribute main_view.
(This requires some delicate timing as to when the controllable is registered,
as it has to happen after the creation of the view in question.)  Right now,
this main_view is used for getting slide change events; it will also be the
one on which controls will be drawn.