Beginning of GUI controls for movies
This is done with a cairooverlay in gstreamer.  When the movie is playing,
the regular frames give us a chance to update the GUI.  When it's paused, we
add a timeout in which we do a seek to the current location in order to
update the GUI.  Things are sorta-kinda working at this revision, but it's
obvious that we need to refactor the separation between the movie and its
controls.  We also need to dive into the internals of pdfpc and change how
it handles button_press_events.